Current User Utils
Would have been nice to have them in KurbSideUtils but you can't access the KSContext, due to dependencies; Attempting to add the KurbSide dependency would result in circular dependencies which is a no-no.
Any ideas @sethwv ? If not, we can move it to a different folder as it doesn't really belong with the services.